Tidy up the file system.
