+++ data/duke.txt
@@ -1,1 +1,2 @@
-1.[E][ ] dad m (from: Dec 23 2023 0000hrs to: Dec 24 2023 0010hrs)
+1.[E][X] dad m (from: Dec 23 2023 0000hrs to: Dec 24 2023 0010hrs)
+2.[T][X] wkdan da
