+++ src/main/java/DukeException.java
@@ -1,0 +1,12 @@
+/**
+ * Encapsulates all exceptions relating to Duke.
+ */
+public class DukeException extends Exception{
+    public DukeException() {
+        super();
+    }
+
+    public DukeException(String message) {
+        super(message);
+    }
+}
